Changed level start camera behavior; Fixed vignette alpha
Camera now starts where it is placed in the scene and swoops toward the
player.
The scenes have been modified so that the camera starts close to the end
tile then swoops toward the player. This might give the player a better
sense of where the end of the level is. (Note camera is hand placed in
all scenes)